closures with parameters (UIControl or UIGestureRecognizer)
The extensions on UIControl and UIGestureRecognizer can be called with a closure with these objects as parameters. They also still can be called with a closure without parameters.

There is an internal enum ActionKitClosure, which contains these 3 types of closures: .NoParameters, .WithControlParameter and .WithGestureParameter. For each of these closures there is an internal typealias.